Pragmatic enforcement fix for orchestrator/subagent separation
- Use mode-based enforcement instead of agentId detection
- Add Bash whitelist for orchestrator (workflow commands, git, gh)
- Add Task output truncation (5000 chars) to prevent context bloat
- Add task retry limits (3 failures before escalation)
- Fix TDD kick-back: test failure→implement, coverage gap→test_writing
- Rename combined-enforcement.py to combined_enforcement.py
- Add 11 e2e workflow tests
