+++ hooks/hooks.json
@@ -7,7 +7,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "python3 ${CLAUDE_PLUGIN_ROOT}/hooks/combined-enforcement.py",
+            "command": "python3 ${CLAUDE_PLUGIN_ROOT}/hooks/combined_enforcement.py",
             "timeout": 3
           }
         ]
