+++ pyproject.toml
@@ -22,7 +22,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
-pythonpath = [".", "scripts", "hooks"]
+pythonpath = [".", "scripts", "hooks", "lib"]
 
 [tool.mypy]
 python_version = "3.10"
